Remove registry mounting on k3s
k3d has supported the `image import` argument for several months. This change is reflected in our documentation,
but we still keep it in the code registry mounting directly in the k3s cluster - for backward compatibility.
With the switch to `@v2` I'm are removing the registry mounting directly in the k3s cluster,
as this functionality is already natively supported in k3d.

Signed-off-by: kuritka <kuritka@gmail.com>