Removing network configuration
Advanced networking is not used anymore as k3d provides native way how to works with networks.
With latest k3d, the network can be specified in the configuration or via argument.
Subnet unfortunately not and I couldn't even use subnet as a separate argument.
I think it would be best to make a PR to k3d to define the subnet in the configuration.

I decided to delete the networking configurations.

Signed-off-by: kuritka <kuritka@gmail.com>